feat: implemented MVVM architecture on MainActivty
All business logic for MainActivity now resides in its view model. MainActivity is only responsible for the view logic, whereas MainActivityViewModel is responsible for the business logic. Binding of the model data from the view model to the views in the main activity is done auto-magically with the aid of the android data binding library